pyproject.toml
[tool.poetry]
description = "CGP_Master_node"
package-mode = false
authors = ["KondratevProkhor <prokhor-2000@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.2"
djangorestframework-gis = "^1.0"
psycopg2-binary = "^2.9.9"
daphne = "^4.1.2"
redis = "^5.0.6"
celery = "^5.4.0"
flower = "^2.0.1"
django-cors-headers = "^4.4.0"
aiohttp = "^3.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
